websockets/constructor/005.html

<!doctype html>
<title>WebSockets: return value</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../constants.js></script>
<div id=log></div>
<script>
var domain = SCHEME_DOMAIN_PORT.split('://')[1];

test(function() {
  assert_true(new WebSocket("ws://"+domain+"/") instanceof WebSocket);
});
</script>
